refactor: rename datocms.d.ts to types.ts to be more consistent
