+++ .eslintignore
@@ -1,5 +1,5 @@
 dist
 node_modules
 
-src/lib/datocms/datocms.d.ts
+src/lib/datocms/types.ts
 !**/.graphqlrc.ts
