+++ .gitignore
@@ -17,7 +17,7 @@ node_modules/
 # generated files
 src/assets/icon-sprite.svg
 src/assets/icon-sprite.d.ts
-src/lib/datocms/datocms.d.ts
+src/lib/datocms/types.ts
 src/lib/i18n/messages.json
 src/lib/i18n/types.ts
 src/lib/site.json
