+++ config/plop/templates/block/Block.astro.hbs
@@ -1,5 +1,5 @@
 ---
-import type { {{ pascalCase name }}Fragment } from '@lib/types/datocms';
+import type { {{ pascalCase name }}Fragment } from '@lib/datocms/types';
 
 export interface Props {
     block: {{ pascalCase name }}Fragment
