+++ config/plop/templates/page/route.astro.hbs
@@ -1,7 +1,7 @@
 ---
 {{#if name }}
 import { datocmsRequest } from '@lib/datocms';
-import type { {{ name }}Query } from '@lib/types/datocms';
+import type { {{ name }}Query } from '@lib/datocms/types';
 import query from './_index.query.graphql';
 {{/if}}
 import Layout from '@layouts/Default.astro';
