+++ src/blocks/EmbedBlock/EmbedBlock.astro
@@ -1,5 +1,5 @@
 ---
-import type { EmbedBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { EmbedBlockFragment } from '@lib/datocms/types';
 import { getOEmbedProvider } from './index';
 import type { OEmbedAny } from './index';
 import BasicEmbed from './embeds/Basic.astro';
