+++ src/blocks/EmbedBlock/embeds/Vimeo.astro
@@ -1,6 +1,6 @@
 ---
 import VideoEmbedBlock from '@blocks/VideoEmbedBlock/VideoEmbedBlock.astro';
-import type { VideoEmbedBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { VideoEmbedBlockFragment } from '@lib/datocms/types';
 import type { OEmbedVideo } from '../index';
 
 type OEmbedVimeo = OEmbedVideo & {
