+++ src/blocks/EmbedBlock/embeds/YouTube.astro
@@ -1,7 +1,7 @@
 ---
 import getVideoId from 'get-video-id';
 import VideoEmbedBlock from '@blocks/VideoEmbedBlock/VideoEmbedBlock.astro';
-import type { VideoEmbedBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { VideoEmbedBlockFragment } from '@lib/datocms/types';
 import type { OEmbedVideo } from '../index';
 
 type OEmbedYouTube = OEmbedVideo & {
