+++ src/blocks/ImageBlock/ImageBlock.astro
@@ -1,5 +1,5 @@
 ---
-import type { ImageBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { ImageBlockFragment } from '@lib/datocms/types';
 import Image from './Image.astro';
 
 export interface Props {
