+++ src/blocks/InternalLink/InternalLink.astro
@@ -3,7 +3,7 @@ import type {
   InternalLinkFragment,
   PageRecord,
   SiteLocale,
-} from '@lib/datocms/datocms.d.ts';
+} from '@lib/datocms/types';
 import { getPagePath } from './index';
 
 export interface Props {
