+++ src/blocks/PagePartialBlock/PagePartialBlock.astro
@@ -1,5 +1,5 @@
 ---
-import type { PagePartialBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { PagePartialBlockFragment } from '@lib/datocms/types';
 import type { AnyBlock } from '@blocks/Blocks';
 import Blocks from '@blocks/Blocks.astro';
 import { Accordion, AccordionItem } from '@components/Accordion';
