+++ src/blocks/TableBlock/TableBlock.astro
@@ -1,5 +1,5 @@
 ---
-import type { TableBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { TableBlockFragment } from '@lib/datocms/types';
 import DynamicTag from '@components/DynamicTag.astro';
 
 type TableRow = {
