+++ src/blocks/TextBlock/TextBlock.astro
@@ -1,6 +1,6 @@
 ---
 import type { StructuredText } from 'datocms-structured-text-utils';
-import type { TextBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { TextBlockFragment } from '@lib/datocms/types';
 import Text from './Text.astro';
 
 export interface Props {
