+++ src/blocks/TextBlock/nodes/ItemLink.astro
@@ -1,6 +1,6 @@
 ---
 import InternalLink from '@blocks/InternalLink/InternalLink.astro';
-import type { InternalLinkFragment } from '@lib/datocms/datocms.d.ts';
+import type { InternalLinkFragment } from '@lib/datocms/types';
 
 type Node = {
   meta?: {
