+++ src/blocks/TextImageBlock/TextImageBlock.astro
@@ -1,6 +1,6 @@
 ---
 import type { StructuredText } from 'datocms-structured-text-utils';
-import type { TextImageBlockFragment } from '@lib/datocms/datocms.d.ts';
+import type { TextImageBlockFragment } from '@lib/datocms/types';
 import Text from '@blocks/TextBlock/Text.astro';
 import Image from '@blocks/ImageBlock/Image.astro';
