+++ src/blocks/VideoBlock/VideoBlock.astro
@@ -1,5 +1,5 @@
 ---
-import type { UploadVideoField, VideoBlockFragment, VideoTextTrackRecord } from '@lib/datocms/datocms.d.ts';
+import type { UploadVideoField, VideoBlockFragment, VideoTextTrackRecord } from '@lib/datocms/types';
 import VideoFallback from './VideoFallback.astro';
 import VideoTextTracks from './VideoTextTracks.astro';
