+++ src/blocks/VideoBlock/VideoFallback.astro
@@ -1,5 +1,5 @@
 ---
-import type { UploadVideoField } from '@lib/datocms/datocms.d.ts';
+import type { UploadVideoField } from '@lib/datocms/types';
 import { t } from '@lib/i18n';
 
 interface Props {
