+++ src/blocks/VideoBlock/VideoTextTracks.astro
@@ -1,5 +1,5 @@
 ---
-import type { VideoTextTrackRecord } from '@lib/datocms/datocms.d.ts';
+import type { VideoTextTrackRecord } from '@lib/datocms/types';
 import { getLocaleName } from '@lib/i18n';
 
 interface Props {
