+++ src/blocks/VideoEmbedBlock/VideoEmbedBlock.astro
@@ -1,5 +1,5 @@
 ---
-import type { VideoEmbedBlockFragment, VideoField } from '@lib/datocms/datocms.d.ts';
+import type { VideoEmbedBlockFragment, VideoField } from '@lib/datocms/types';
 import { t } from '@lib/i18n';
 import Icon from '@components/Icon';
 import PlayButton from './PlayButton.astro';
