+++ src/components/LocaleSelector/LocaleSelector.astro
@@ -1,6 +1,6 @@
 ---
 import { getLocale, getLocaleName, locales, t } from '@lib/i18n';
-import type { SiteLocale } from '@lib/datocms/datocms.d.ts';
+import type { SiteLocale } from '@lib/datocms/types';
 import type { PageUrl } from '@lib/seo';
 import '@assets/a11y.css';
