+++ src/components/SearchResult.astro
@@ -1,6 +1,6 @@
 ---
 import DynamicTag from './DynamicTag.astro';
-import type { Tag } from '@lib/datocms/datocms.d.ts';
+import type { Tag } from '@lib/datocms/types';
 
 type SearchResult = {
   title: string
