+++ src/components/SeoHead.astro
@@ -3,7 +3,7 @@ import DynamicTag from './DynamicTag.astro';
 import { getLocale } from '@lib/i18n';
 import { getOpenSearchName, getOpenSearchPathname } from '@lib/search';
 import type { PageUrl } from '@lib/seo';
-import type { Tag } from '@lib/datocms/datocms.d.ts';
+import type { Tag } from '@lib/datocms/types';
 
 interface Props {
   pageUrls: PageUrl[];
