+++ src/layouts/Default.astro
@@ -1,5 +1,5 @@
 ---
-import type { DefaultLayoutQuery, Site, SiteLocale, Tag } from '@lib/datocms/datocms.d.ts';
+import type { DefaultLayoutQuery, Site, SiteLocale, Tag } from '@lib/datocms/types';
 import { datocmsRequest } from '@lib/datocms';
 import { defaultLocale, t } from '@lib/i18n';
 import { datocmsNoIndex, getSearchPathname } from '@lib/search';
