+++ src/lib/seo/index.ts
@@ -1,4 +1,4 @@
-import type { Tag } from '@lib/datocms/datocms';
+import type { Tag } from '@lib/datocms/types';
 import { getLocale } from '@lib/i18n';
 import type { SiteLocale } from '@lib/i18n/types';
 import { globalSeo } from '@lib/site.json';
