+++ src/pages/404.astro
@@ -3,7 +3,7 @@ import type { SiteLocale } from '@lib/i18n/types';
 import type {
   NotFoundPageQuery,
   NotFoundPageRecord,
-} from '@lib/datocms/datocms.d.ts';
+} from '@lib/datocms/types';
 import { datocmsRequest } from '@lib/datocms';
 import { noIndexTag, titleTag } from '@lib/seo';
 import Layout from '@layouts/Default.astro';
