+++ src/pages/[locale]/[...path]/index.astro
@@ -1,6 +1,6 @@
 ---
 import { datocmsCollection, datocmsRequest } from '@lib/datocms';
-import type { PageQuery, PageRecord, ParentPageFragment, SiteLocale } from '@lib/datocms/datocms.d.ts';
+import type { PageQuery, PageRecord, ParentPageFragment, SiteLocale } from '@lib/datocms/types';
 import type { PageUrl } from '@lib/seo';
 import Layout from '@layouts/Default.astro';
 import Blocks from '@blocks/Blocks.astro';
