+++ src/pages/[locale]/index.astro
@@ -1,6 +1,6 @@
 ---
 import { datocmsRequest } from '@lib/datocms';
-import type { HomePageQuery, HomePageRecord } from '@lib/datocms/datocms.d.ts';
+import type { HomePageQuery, HomePageRecord } from '@lib/datocms/types';
 import type { PageUrl } from '@lib/seo';
 import { locales } from '@lib/i18n';
 import Layout from '@layouts/Default.astro';
